+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release Notes - heidelpay Payment Gateway for WooCommerce
 
+## 1.1.0
+
+### Added
+
+#### New payment methods:
+ - iDeal
+ - PayPal
+ - Direct Debit
+
+#### Features:
+- Credit Card and Debit Card: Added css file to customize the appearance of the iFrame.
+- Credit Card, Debit Card and Paypal: Added setting to chose between bookingmodes debit (default) and authorize.
+
+### Changed
+
+
 ## 1.0.1
 
 ### Fixed